Moved pet helpers into a separate file
