+++ src/components/Pet.tsx
@@ -1,16 +1,4 @@
-export enum PetMood {
-  HAPPY,
-  NEUTRAL,
-  SAD,
-  EXCITED,
-  SLEEPING,
-  DEAD,
-}
-
-/**
- * Each frame takes up 100ms, and most of the animations take up 6 frames. So the animation time should be 600ms. The only exceptions are the "sleeping" animation which is 18 frames and the "dead" art which is 1 frame.
- */
-export const ANIMATION_TIME = 600;
+import { getPetAnimation, PetMood } from "@/types/pet";
 
 interface Props {
   mood: PetMood;
@@ -31,34 +19,4 @@ export const Pet = (props: Props) => {
       }}
     />
   );
-};
-
-export const getPetAnimation = (mood: PetMood) => {
-  const filePath = "/gifs/";
-  let fileName = "";
-
-  switch (mood) {
-    case PetMood.HAPPY:
-      fileName = "happy.gif";
-      break;
-    case PetMood.NEUTRAL:
-      fileName = "neutral.gif";
-      break;
-    case PetMood.SAD:
-      fileName = "sad.gif";
-      break;
-    case PetMood.EXCITED:
-      fileName = "excited.gif";
-      break;
-    case PetMood.SLEEPING:
-      fileName = "sleeping.gif";
-      break;
-    case PetMood.DEAD:
-      fileName = "dead.png";
-      break;
-    default:
-      fileName = "happy.gif";
-  }
-
-  return filePath + fileName;
 };
