+++ src/components/PetCreationForm.tsx
@@ -1,9 +1,10 @@
-import { Pet, PetMood } from "@/components/Pet";
+import { Pet } from "@/components/Pet";
 import { Box, Button, Stack, TextField, Typography } from "@mui/material";
 import { useMutation } from "convex/react";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
 import { api } from "../../convex/_generated/api";
+import { PetMood } from "@/types/pet";
 
 interface PetCreationFormData {
   petName: string;
